+++ qoqo/src/measurements/basis_rotation_measurement.rs
@@ -14,7 +14,7 @@
 
 use super::PauliZProductInputWrapper;
 use crate::CircuitWrapper;
-use bincode::serialize;
+use bincode::{deserialize, serialize};
 use pyo3::exceptions::{PyRuntimeError, PyTypeError, PyValueError};
 use pyo3::prelude::*;
 use pyo3::types::PyByteArray;
@@ -222,6 +222,46 @@ impl PauliZProductWrapper {
         Ok(("PauliZProduct", b))
     }
 
+    /// Return the bincode representation of the PauliZProduct using the [bincode] crate.
+    ///
+    /// Returns:
+    ///     ByteArray: The serialized PauliZProduct (in [bincode] form).
+    ///
+    /// Raises:
+    ///     ValueError: Cannot serialize PauliZProduct to bytes.
+    pub fn to_bincode(&self) -> PyResult<Py<PyByteArray>> {
+        let serialized = serialize(&self.internal)
+            .map_err(|_| PyValueError::new_err("Cannot serialize PauliZProduct to bytes"))?;
+        let b: Py<PyByteArray> = Python::with_gil(|py| -> Py<PyByteArray> {
+            PyByteArray::new(py, &serialized[..]).into()
+        });
+        Ok(b)
+    }
+
+    #[staticmethod]
+    /// Convert the bincode representation of the PauliZProduct to a PauliZProduct using the [bincode] crate.
+    ///
+    /// Args:
+    ///     input (ByteArray): The serialized PauliZProduct (in [bincode] form).
+    ///
+    /// Returns:
+    ///     PauliZProduct: The deserialized PauliZProduct.
+    ///
+    /// Raises:
+    ///     TypeError: Input cannot be converted to byte array.
+    ///     ValueError: Input cannot be deserialized to PauliZProduct.
+    pub fn from_bincode(input: &PyAny) -> PyResult<Self> {
+        let bytes = input
+            .extract::<Vec<u8>>()
+            .map_err(|_| PyTypeError::new_err("Input cannot be converted to byte array"))?;
+
+        Ok(Self {
+            internal: deserialize(&bytes[..]).map_err(|_| {
+                PyValueError::new_err("Input cannot be deserialized to PauliZProduct")
+            })?,
+        })
+    }
+
     /// Serialize the PauliZProduct to json form using the [serde_json] crate.
     ///
     /// Returns:
@@ -287,5 +327,38 @@ impl PauliZProductWrapper {
                 "Other comparison not implemented",
             )),
         }
+    }
+}
+
+impl PauliZProductWrapper {
+    /// Extracts a PauliZProduct from a PauliZProductWrapper python object.
+    ///
+    /// When working with qoqo and other rust based python packages compiled separately
+    /// a downcast will not detect that two PauliZProductWrapper objects are compatible.
+    /// Provides a custom function to convert qoqo PauliZProducts between different Python packages.
+    ///
+    /// # Arguments:
+    ///
+    /// `input` - The Python object that should be casted to a [roqoqo::PauliZProduct]
+    pub fn from_pyany(input: Py<PyAny>) -> PyResult<PauliZProduct> {
+        Python::with_gil(|py| -> PyResult<PauliZProduct> {
+            let input = input.as_ref(py);
+            if let Ok(try_downcast) = input.extract::<PauliZProductWrapper>() {
+                Ok(try_downcast.internal)
+            } else {
+                let get_bytes = input.call_method0("to_bincode").map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo PauliZProduct: Cast to binary representation failed".to_string())
+            })?;
+                let bytes = get_bytes.extract::<Vec<u8>>().map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo PauliZProduct: Cast to binary representation failed".to_string())
+            })?;
+                deserialize(&bytes[..]).map_err(|err| {
+                    PyTypeError::new_err(format!(
+                    "Python object cannot be converted to qoqo PauliZProduct: Deserialization failed: {}",
+                    err
+                ))
+                })
+            }
+        })
     }
 }