+++ qoqo/src/measurements/cheated_basis_rotation_measurement.rs
@@ -14,7 +14,7 @@
 
 use super::CheatedPauliZProductInputWrapper;
 use crate::CircuitWrapper;
-use bincode::serialize;
+use bincode::{deserialize, serialize};
 use pyo3::exceptions::{PyRuntimeError, PyTypeError, PyValueError};
 use pyo3::prelude::*;
 use pyo3::types::PyByteArray;
@@ -221,6 +221,46 @@ impl CheatedPauliZProductWrapper {
         Ok(("CheatedPauliZProduct", b))
     }
 
+    /// Return the bincode representation of the CheatedPauliZProduct using the [bincode] crate.
+    ///
+    /// Returns:
+    ///     ByteArray: The serialized CheatedPauliZProduct (in [bincode] form).
+    ///
+    /// Raises:
+    ///     ValueError: Cannot serialize CheatedPauliZProduct to bytes.
+    pub fn to_bincode(&self) -> PyResult<Py<PyByteArray>> {
+        let serialized = serialize(&self.internal)
+            .map_err(|_| PyValueError::new_err("Cannot serialize CheatedPauliZProduct to bytes"))?;
+        let b: Py<PyByteArray> = Python::with_gil(|py| -> Py<PyByteArray> {
+            PyByteArray::new(py, &serialized[..]).into()
+        });
+        Ok(b)
+    }
+
+    #[staticmethod]
+    /// Convert the bincode representation of the CheatedPauliZProduct to a CheatedPauliZProduct using the [bincode] crate.
+    ///
+    /// Args:
+    ///     input (ByteArray): The serialized CheatedPauliZProduct (in [bincode] form).
+    ///
+    /// Returns:
+    ///     CheatedPauliZProduct: The deserialized CheatedPauliZProduct.
+    ///
+    /// Raises:
+    ///     TypeError: Input cannot be converted to byte array.
+    ///     ValueError: Input cannot be deserialized to CheatedPauliZProduct.
+    pub fn from_bincode(input: &PyAny) -> PyResult<Self> {
+        let bytes = input
+            .extract::<Vec<u8>>()
+            .map_err(|_| PyTypeError::new_err("Input cannot be converted to byte array"))?;
+
+        Ok(Self {
+            internal: deserialize(&bytes[..]).map_err(|_| {
+                PyValueError::new_err("Input cannot be deserialized to CheatedPauliZProduct")
+            })?,
+        })
+    }
+
     /// Serializes the CheatedPauliZProduct to json form using the [serde_json] crate.
     ///
     /// Returns:
@@ -286,5 +326,38 @@ impl CheatedPauliZProductWrapper {
                 "Other comparison not implemented",
             )),
         }
+    }
+}
+
+impl CheatedPauliZProductWrapper {
+    /// Extracts a CheatedPauliZProduct from a CheatedPauliZProductWrapper python object.
+    ///
+    /// When working with qoqo and other rust based python packages compiled separately
+    /// a downcast will not detect that two CheatedPauliZProductWrapper objects are compatible.
+    /// Provides a custom function to convert qoqo CheatedPauliZProducts between different Python packages.
+    ///
+    /// # Arguments:
+    ///
+    /// `input` - The Python object that should be casted to a [roqoqo::CheatedPauliZProduct]
+    pub fn from_pyany(input: Py<PyAny>) -> PyResult<CheatedPauliZProduct> {
+        Python::with_gil(|py| -> PyResult<CheatedPauliZProduct> {
+            let input = input.as_ref(py);
+            if let Ok(try_downcast) = input.extract::<CheatedPauliZProductWrapper>() {
+                Ok(try_downcast.internal)
+            } else {
+                let get_bytes = input.call_method0("to_bincode").map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo CheatedPauliZProduct: Cast to binary representation failed".to_string())
+            })?;
+                let bytes = get_bytes.extract::<Vec<u8>>().map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo CheatedPauliZProduct: Cast to binary representation failed".to_string())
+            })?;
+                deserialize(&bytes[..]).map_err(|err| {
+                    PyTypeError::new_err(format!(
+                    "Python object cannot be converted to qoqo CheatedPauliZProduct: Deserialization failed: {}",
+                    err
+                ))
+                })
+            }
+        })
     }
 }