+++ qoqo/src/measurements/cheated_measurement.rs
@@ -14,7 +14,7 @@
 
 use super::CheatedInputWrapper;
 use crate::CircuitWrapper;
-use bincode::serialize;
+use bincode::{deserialize, serialize};
 use pyo3::exceptions::{PyRuntimeError, PyTypeError, PyValueError};
 use pyo3::prelude::*;
 use pyo3::types::PyByteArray;
@@ -221,6 +221,45 @@ impl CheatedWrapper {
         Ok(("Cheated", b))
     }
 
+    /// Return the bincode representation of the Cheated using the [bincode] crate.
+    ///
+    /// Returns:
+    ///     ByteArray: The serialized Cheated (in [bincode] form).
+    ///
+    /// Raises:
+    ///     ValueError: Cannot serialize Cheated to bytes.
+    pub fn to_bincode(&self) -> PyResult<Py<PyByteArray>> {
+        let serialized = serialize(&self.internal)
+            .map_err(|_| PyValueError::new_err("Cannot serialize Cheated to bytes"))?;
+        let b: Py<PyByteArray> = Python::with_gil(|py| -> Py<PyByteArray> {
+            PyByteArray::new(py, &serialized[..]).into()
+        });
+        Ok(b)
+    }
+
+    #[staticmethod]
+    /// Convert the bincode representation of the Cheated to a Cheated using the [bincode] crate.
+    ///
+    /// Args:
+    ///     input (ByteArray): The serialized Cheated (in [bincode] form).
+    ///
+    /// Returns:
+    ///     Cheated: The deserialized Cheated.
+    ///
+    /// Raises:
+    ///     TypeError: Input cannot be converted to byte array.
+    ///     ValueError: Input cannot be deserialized to Cheated.
+    pub fn from_bincode(input: &PyAny) -> PyResult<Self> {
+        let bytes = input
+            .extract::<Vec<u8>>()
+            .map_err(|_| PyTypeError::new_err("Input cannot be converted to byte array"))?;
+
+        Ok(Self {
+            internal: deserialize(&bytes[..])
+                .map_err(|_| PyValueError::new_err("Input cannot be deserialized to Cheated"))?,
+        })
+    }
+
     /// Serialize the Cheated measurement to json form.
     ///
     /// Returns:
@@ -286,5 +325,38 @@ impl CheatedWrapper {
                 "Other comparison not implemented",
             )),
         }
+    }
+}
+
+impl CheatedWrapper {
+    /// Extracts a Cheated from a CheatedWrapper python object.
+    ///
+    /// When working with qoqo and other rust based python packages compiled separately
+    /// a downcast will not detect that two CheatedWrapper objects are compatible.
+    /// Provides a custom function to convert qoqo Cheateds between different Python packages.
+    ///
+    /// # Arguments:
+    ///
+    /// `input` - The Python object that should be casted to a [roqoqo::Cheated]
+    pub fn from_pyany(input: Py<PyAny>) -> PyResult<Cheated> {
+        Python::with_gil(|py| -> PyResult<Cheated> {
+            let input = input.as_ref(py);
+            if let Ok(try_downcast) = input.extract::<CheatedWrapper>() {
+                Ok(try_downcast.internal)
+            } else {
+                let get_bytes = input.call_method0("to_bincode").map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo Cheated: Cast to binary representation failed".to_string())
+            })?;
+                let bytes = get_bytes.extract::<Vec<u8>>().map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo Cheated: Cast to binary representation failed".to_string())
+            })?;
+                deserialize(&bytes[..]).map_err(|err| {
+                    PyTypeError::new_err(format!(
+                    "Python object cannot be converted to qoqo Cheated: Deserialization failed: {}",
+                    err
+                ))
+                })
+            }
+        })
     }
 }