+++ qoqo/src/measurements/classical_register_measurement.rs
@@ -13,7 +13,7 @@
 //! Qoqo classical registers
 
 use crate::CircuitWrapper;
-use bincode::serialize;
+use bincode::{deserialize, serialize};
 use pyo3::exceptions::{PyRuntimeError, PyTypeError, PyValueError};
 use pyo3::prelude::*;
 use pyo3::types::PyByteArray;
@@ -143,6 +143,46 @@ impl ClassicalRegisterWrapper {
         Ok(("ClassicalRegister", b))
     }
 
+    /// Return the bincode representation of the ClassicalRegister using the [bincode] crate.
+    ///
+    /// Returns:
+    ///     ByteArray: The serialized ClassicalRegister (in [bincode] form).
+    ///
+    /// Raises:
+    ///     ValueError: Cannot serialize ClassicalRegister to bytes.
+    pub fn to_bincode(&self) -> PyResult<Py<PyByteArray>> {
+        let serialized = serialize(&self.internal)
+            .map_err(|_| PyValueError::new_err("Cannot serialize ClassicalRegister to bytes"))?;
+        let b: Py<PyByteArray> = Python::with_gil(|py| -> Py<PyByteArray> {
+            PyByteArray::new(py, &serialized[..]).into()
+        });
+        Ok(b)
+    }
+
+    #[staticmethod]
+    /// Convert the bincode representation of the ClassicalRegister to a ClassicalRegister using the [bincode] crate.
+    ///
+    /// Args:
+    ///     input (ByteArray): The serialized ClassicalRegister (in [bincode] form).
+    ///
+    /// Returns:
+    ///     ClassicalRegister: The deserialized ClassicalRegister.
+    ///
+    /// Raises:
+    ///     TypeError: Input cannot be converted to byte array.
+    ///     ValueError: Input cannot be deserialized to ClassicalRegister.
+    pub fn from_bincode(input: &PyAny) -> PyResult<Self> {
+        let bytes = input
+            .extract::<Vec<u8>>()
+            .map_err(|_| PyTypeError::new_err("Input cannot be converted to byte array"))?;
+
+        Ok(Self {
+            internal: deserialize(&bytes[..]).map_err(|_| {
+                PyValueError::new_err("Input cannot be deserialized to ClassicalRegister")
+            })?,
+        })
+    }
+
     /// Serialize the ClassicalRegister measurement to json form.
     ///
     /// Returns:
@@ -208,5 +248,38 @@ impl ClassicalRegisterWrapper {
                 "Other comparison not implemented",
             )),
         }
+    }
+}
+
+impl ClassicalRegisterWrapper {
+    /// Extracts a ClassicalRegister from a ClassicalRegisterWrapper python object.
+    ///
+    /// When working with qoqo and other rust based python packages compiled separately
+    /// a downcast will not detect that two ClassicalRegisterWrapper objects are compatible.
+    /// Provides a custom function to convert qoqo ClassicalRegisters between different Python packages.
+    ///
+    /// # Arguments:
+    ///
+    /// `input` - The Python object that should be casted to a [roqoqo::ClassicalRegister]
+    pub fn from_pyany(input: Py<PyAny>) -> PyResult<ClassicalRegister> {
+        Python::with_gil(|py| -> PyResult<ClassicalRegister> {
+            let input = input.as_ref(py);
+            if let Ok(try_downcast) = input.extract::<ClassicalRegisterWrapper>() {
+                Ok(try_downcast.internal)
+            } else {
+                let get_bytes = input.call_method0("to_bincode").map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo ClassicalRegister: Cast to binary representation failed".to_string())
+            })?;
+                let bytes = get_bytes.extract::<Vec<u8>>().map_err(|_| {
+                PyTypeError::new_err("Python object cannot be converted to qoqo ClassicalRegister: Cast to binary representation failed".to_string())
+            })?;
+                deserialize(&bytes[..]).map_err(|err| {
+                    PyTypeError::new_err(format!(
+                    "Python object cannot be converted to qoqo ClassicalRegister: Deserialization failed: {}",
+                    err
+                ))
+                })
+            }
+        })
     }
 }