+++ qoqo/tests/integration/quantum_program.rs
@@ -447,7 +447,7 @@ fn test_to_from_bincode() {
         let serialised_error = serialised.call_method0("to_bincode");
         assert!(serialised_error.is_err());
 
-        // testing that 'from_bincode' can be called directly on a QuantumProgram (python classmethod)
+        // testing that 'from_bincode' can be called directly on a QuantumProgram (python staticmethod)
         let deserialised_py = program_type
             .call_method1("from_bincode", (serialised,))
             .unwrap();
@@ -527,7 +527,7 @@ fn test_to_from_json() {
         let deserialised_error = deserialised.call_method0("from_json");
         assert!(deserialised_error.is_err());
 
-        // testing that 'from_json' can be called directly on a QuantumProgram (python classmethod)
+        // testing that 'from_json' can be called directly on a QuantumProgram (python staticmethod)
         let deserialised_py = program_type
             .call_method1("from_json", (serialised,))
             .unwrap();